requirements.txt
streamlit==1.33.0
pypdf2==3.0.1
langchain==0.1.16
sentence-transformers==2.7.0
transformers==4.40.1
tokenizers==0.19.1
faiss-cpu==1.8.0
google-generativeai==0.5.4
python-dotenv==1.0.1
numpy==1.26.4